client: move tests that rely on Dial to a separate file
